requirements.txt
absl-py==1.0.0
altgraph==0.17.2
attrs==21.4.0
auto-py-to-exe==2.19.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cycler==0.11.0
distlib==0.3.4
docopt==0.6.2
Eel==0.14.0
filelock==3.7.0
fonttools==4.33.3
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
keyboard==0.13.5
kiwisolver==1.4.2
matplotlib==3.5.2
mediapipe==0.8.10
Nuitka==0.8.1
numpy==1.22.3
opencv-contrib-python==4.5.5.64
packaging==21.3
pefile==2021.9.3
Pillow==9.1.0
pipreqs==0.4.11
platformdirs==2.5.2
protobuf==3.20.1
pycparser==2.21
pygame==2.1.2
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.5
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pywin32-ctypes==0.2.0
requests==2.27.1
six==1.16.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
whichcraft==0.6.1
yarg==0.1.9
zope.event==4.5.0
zope.interface==5.4.0
